[backend] fix(ts): assert non-null/undefined on DB data returned

+++ apps/backend/src/modules/deployment/deployment.app.ts
@@ -474,7 +474,10 @@ export const DeploymentApp = {
       ![
         DeploymentRequestPlatformState.Unprovisioned,
         DeploymentRequestPlatformState.Provisioning,
-      ].includes(deploymentRequest.actual_state);
+      ].includes(
+        deploymentRequest.actual_state ??
+          DeploymentRequestPlatformState.Unprovisioned
+      );
 
     const target_state =
       deploymentRequest.actual_state ===
@@ -549,10 +552,17 @@ export const DeploymentApp = {
     }
 
     try {
-      await auth0Client.deleteAudienceAPI(
-        deploymentRequest.organization_requester_id,
-        deploymentRequest.platform_id
-      );
+      if (deploymentRequest.platform_id) {
+        await auth0Client.deleteAudienceAPI(
+          deploymentRequest.organization_requester_id,
+          deploymentRequest.platform_id
+        );
+      } else {
+        logApp.error('Unable to delete audience', {
+          error: 'missing platform_id',
+          deploymentRequestId: deploymentRequest.id,
+        });
+      }
     } catch (error) {
       logApp.error('Unable to delete audience', {
         error,
@@ -630,10 +640,17 @@ export const DeploymentApp = {
         }
 
         try {
-          await auth0Client.deleteAudienceAPI(
-            trial.organization_requester_id,
-            trial.platform_id
-          );
+          if (trial.platform_id) {
+            await auth0Client.deleteAudienceAPI(
+              trial.organization_requester_id,
+              trial.platform_id
+            );
+          } else {
+            logApp.error('Unable to delete audience', {
+              error: 'missing platform_id',
+              deploymentRequestId: trial.id,
+            });
+          }
         } catch (error) {
           logApp.error('Unable to delete audience', {
             error,
@@ -964,10 +981,10 @@ const sendUpdateDeploymentTelemetryEvent = async (
         deployment_id: deploymentRequest.id,
         deployment_type: deploymentRequest.type,
         platform_id: deploymentRequest.platform_id,
-        ...(deploymentRequest.hub_status ===
-          DeploymentRequestHubStatus.Cancelled && {
-          cancellation_reason: deploymentRequest.cancellation_reason,
-        }),
+        cancellation_reason:
+          deploymentRequest.hub_status === DeploymentRequestHubStatus.Cancelled
+            ? deploymentRequest.cancellation_reason
+            : undefined,
       }
     );
